Rename embulk-plugin-sqlite3 to embulk-output-sqlite3
