+++ README.md
@@ -1,28 +1,31 @@
-# SQLite3 plugin for Embulk
+# Sqlite3 output plugin for Embulk
 
-embulk-plugin-sqlite3 provides output plugin for Embulk
+TODO: Write short description here and embulk-output-sqlite3.gemspec file.
 
-## Installation
+## Overview
 
-    $ embulk gem install embulk-plugin-sqlite3
+* **Plugin type**: output
+* **Load all or nothing**: no
+* **Resume supported**: no
+* **Cleanup supported**: yes
 
 ## Configuration
+
+- **property1**: description (string, required)
+- **property2**: description (integer, default: default-value)
+
+## Example
 
 ```yaml
 out:
   type: sqlite3
-  database: '/tmp/test.db'
-  table: 'load01'
+  property1: example1
+  property2: example2
 ```
 
-## TODO
 
-- implement input pugin
+## Build
 
-## Contributing
-
-1. Fork it ( https://github.com/takady/embulk-plugin-sqlite3/fork )
-2. Create your feature branch (`git checkout -b my-new-feature`)
-3. Commit your changes (`git commit -am 'Add some feature'`)
-4. Push to the branch (`git push origin my-new-feature`)
-5. Create a new Pull Request
+```
+$ rake
+```
